chore: update source to creek 0.4.4-SNAPSHOT API (package restructure)
Creek 0.4.4-SNAPSHOT moved topic classes from org.creekservice.api.kafka.metadata
to org.creekservice.api.kafka.metadata.topic, and PartDescriptor gained
new name() and topic() methods.
