+++ services/src/main/java/org/acme/example/services/ExampleServiceDescriptor.java
@@ -23,8 +23,8 @@ import java.util.Collection;
 import java.util.List;
 import org.acme.example.api.ExampleAggregateDescriptor;
 import org.acme.example.internal.TopicConfigBuilder;
-import org.creekservice.api.kafka.metadata.OwnedKafkaTopicInput;
-import org.creekservice.api.kafka.metadata.OwnedKafkaTopicOutput;
+import org.creekservice.api.kafka.metadata.topic.OwnedKafkaTopicInput;
+import org.creekservice.api.kafka.metadata.topic.OwnedKafkaTopicOutput;
 import org.creekservice.api.platform.metadata.ComponentInput;
 import org.creekservice.api.platform.metadata.ComponentInternal;
 import org.creekservice.api.platform.metadata.ComponentOutput;
